transitions and code cleaning
